Refactor code and moved into gora-jet module
