--- gora-hazelcast-jet/src/main/avro/metricdatum.json
@@ -1,10 +1,0 @@
-{
-  "type": "record",
-  "name": "MetricDatum", "default":null,
-  "namespace": "org.apache.gora.hazelcastJet.generated",
-  "fields" : [
-    {"name": "metricDimension", "type": ["null","string"], "default":null},
-    {"name": "timestamp", "type": "long","default":0},
-    {"name": "metric", "type" : "long","default":0}
-  ]
-}
