--- gora-hazelcast-jet/src/main/java/org/apache/gora/hazelcastJet/JetOutputFormat.java
@@ -1,29 +1,0 @@
-package org.apache.gora.hazelcastJet;
-
-import org.apache.gora.persistency.impl.PersistentBase;
-
-public class JetOutputFormat<KeyOut, ValueOut extends PersistentBase> {
-    public KeyOut key;
-    public ValueOut value;
-
-    public JetOutputFormat(KeyOut key, ValueOut value) {
-        this.key = key;
-        this.value = value;
-    }
-
-    public KeyOut getKey() {
-        return key;
-    }
-
-    public void setKey(KeyOut key) {
-        this.key = key;
-    }
-
-    public ValueOut getValue() {
-        return value;
-    }
-
-    public void setValue(ValueOut value) {
-        this.value = value;
-    }
-}
