--- gora-hazelcast-jet/src/test/avro/ResultPageView.json
@@ -1,10 +1,0 @@
-{
-  "type": "record",
-  "name": "ResultPageView", "default":null,
-  "namespace": "org.apache.gora.hazelcastJet.generated",
-  "fields" : [
-    {"name": "url", "type": ["null","string"], "default":null},
-    {"name": "timestamp", "type": "long", "default":0},
-    {"name": "ip", "type": ["null","string"], "default":null}
-  ]
-}
