--- gora-hazelcast-jet/src/test/avro/pageview.json
@@ -1,15 +1,0 @@
-{
-  "type": "record",
-  "name": "Pageview", "default":null,
-  "namespace": "org.apache.gora.hazelcastJet.generated",
-  "fields" : [
-    {"name": "url", "type": ["null","string"], "default":null},
-    {"name": "timestamp", "type": "long", "default":0},
-    {"name": "ip", "type": ["null","string"], "default":null},
-    {"name": "httpMethod", "type": ["null","string"], "default":null},
-    {"name": "httpStatusCode", "type": "int", "default":0},
-    {"name": "responseSize", "type": "int", "default":0},
-    {"name": "referrer", "type": ["null","string"], "default":null},
-    {"name": "userAgent", "type": ["null","string"], "default":null}
-  ]
-}
