--- gora-hazelcast-jet/src/test/resources/gora-hbase-mapping.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-   Licensed to the Apache Software Foundation (ASF) under one or more
-   contributor license agreements.  See the NOTICE file distributed with
-   this work for additional information regarding copyright ownership.
-   The ASF licenses this file to You under the Apache License, Version 2.0
-   (the "License"); you may not use this file except in compliance with
-   the License.  You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
--->
-
-<!--
-  Gora Mapping file for HBase Backend
--->
-<gora-otd>
-  <table name="Pageview"> <!-- optional descriptors for tables -->
-    <family name="common"/> <!-- This can also have params like compression, bloom filters -->
-    <family name="http"/>
-    <family name="misc"/>
-  </table>
-
-  <table name="ResultPageView"> <!-- optional descriptors for tables -->
-    <family name="common"/> <!-- This can also have params like compression, bloom filters -->
-  </table>
-
-  <class name="org.apache.gora.hazelcastJet.generated.Pageview" keyClass="java.lang.Long" table="AccessLog">
-    <field name="url" family="common" qualifier="url"/>
-    <field name="timestamp" family="common" qualifier="timestamp"/>
-    <field name="ip" family="common" qualifier="ip" />
-    <field name="httpMethod" family="http" qualifier="httpMethod"/>
-    <field name="httpStatusCode" family="http" qualifier="httpStatusCode"/>
-    <field name="responseSize" family="http" qualifier="responseSize"/>
-    <field name="referrer" family="misc" qualifier="referrer"/>
-    <field name="userAgent" family="misc" qualifier="userAgent"/>
-  </class>
-
-  <class name="org.apache.gora.hazelcastJet.generated.MetricDatum" keyClass="java.lang.String" table="Metrics">
-    <field name="metricDimension" family="common"  qualifier="metricDimension"/>
-    <field name="timestamp" family="common" qualifier="ts"/>
-    <field name="metric" family="common" qualifier="metric"/>
-  </class>
-
-  <class name="org.apache.gora.hazelcastJet.generated.ResultPageView" keyClass="java.lang.Long" table="Results">
-    <field name="url" family="common" qualifier="url"/>
-    <field name="timestamp" family="common" qualifier="timestamp"/>
-    <field name="ip" family="common" qualifier="ip" />
-  </class>
-
-</gora-otd>
