--- gora-hazelcast-jet/src/test/resources/test-delete-map-values-addvals.csv
@@ -1,1 +1,0 @@
-[k2#v2,k3#v3]
