--- gora-hazelcast-jet/src/test/resources/test-delete-map-values.csv
@@ -1,1 +1,0 @@
-key1|[k3#v9]
