--- gora-hazelcast-jet/src/test/resources/test-delete-rows.csv
@@ -1,1 +1,0 @@
-key1
