+++ gora-jet/conf/gora-aerospike-mapping.xml
