+++ gora-jet/conf/gora-cassandra-mapping.xml
