+++ gora-jet/conf/gora-couchdb-mapping.xml
