+++ gora-jet/conf/gora-solr-mapping.xml
