+++ gora-jet/src/main/java/org/apache/gora/jet/JetEngine.java
@@ -1,0 +1,26 @@
+package org.apache.gora.jet;
+
+import com.hazelcast.jet.pipeline.BatchSource;
+import com.hazelcast.jet.pipeline.Sink;
+import com.hazelcast.jet.pipeline.Sinks;
+import com.hazelcast.jet.pipeline.Sources;
+import org.apache.gora.persistency.impl.PersistentBase;
+import org.apache.gora.query.Query;
+import org.apache.gora.store.DataStore;
+
+public class JetEngine<KeyIn, ValueIn extends PersistentBase, KeyOut, ValueOut extends PersistentBase> {
+    public static DataStore dataOutStore;
+
+    public BatchSource<ValueIn> createDataSource(Query<KeyIn, ValueIn> query) {
+        BatchSource<ValueIn> source = Sources.batchFromProcessor("gora-jet-source",
+                new JetSource<KeyIn, ValueIn>(query));
+        return source;
+    }
+
+    public Sink<JetOutputFormat<KeyOut, ValueOut>> createDataSink(DataStore<KeyOut, ValueOut> dataOutStore) {
+        JetEngine.dataOutStore = dataOutStore;
+        Sink<JetOutputFormat<KeyOut, ValueOut>> sink = Sinks.fromProcessor("gora-jet-sink",
+                new JetSink<KeyOut, ValueOut>());
+        return sink;
+    }
+}
