+++ gora-jet/src/test/avro/resultPageView.json
@@ -1,0 +1,10 @@
+{
+  "type": "record",
+  "name": "ResultPageView", "default":null,
+  "namespace": "org.apache.gora.jet.generated",
+  "fields" : [
+    {"name": "url", "type": ["null","string"], "default":null},
+    {"name": "timestamp", "type": "long", "default":0},
+    {"name": "ip", "type": ["null","string"], "default":null}
+  ]
+}
