+++ pom.xml
@@ -792,7 +792,7 @@
     <module>gora-ignite</module>
     <module>gora-tutorial</module>
     <module>sources-dist</module>
-    <module>gora-hazelcast-jet</module>
+    <module>gora-jet</module>
   </modules>
 
   <properties>
